+++ Cargo.toml
@@ -5,8 +5,9 @@ authors = ["Helium <info@helium.com>"]
 edition = "2018"
 
 [features]
-default= []
-services=["tonic-build", "tonic"]
+default = []
+services = ["tonic-build", "tonic"]
+json_serde = ["byteorder","base64", "bs58"]
 
 [dependencies]
 tonic = {version = "0", optional = true}
@@ -14,9 +15,9 @@ bytes = "1"
 prost = "0"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-base64 = "0"
-bs58 = {version = "0", features=["check"]}
-byteorder = "0"
+base64 = {version = "0", optional = true}
+bs58 = {version = "0", features=["check"], optional = true}
+byteorder =  {version = "0", optional = true}
 
 [build-dependencies]
 tonic-build = {version = "0", optional = true}
